backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
nltk==3.8.1
textblob==0.15.3
numpy==2.2.4
PyJWT==2.8.0
passlib==1.7.4
bcrypt==3.2.0 